Add Spring Boot support
Provide support classes and auto-configuration to allow Atomikos to be
used with Spring Boot. This commit introduces two user-facing modules
along with an integration test module that doesn't need to be published.

The `transactions-spring-boot` class provides support classes along
with auto-configuration. Most of the code was originally developed in
Spring Boot, however, going forward it will be maintained directly here.

The `transactions-spring-boot-starter` provides a starter POM equivalent
to the `spring-boot-starter-jta-atomikos` POM that ships with Spring
Boot 2.3.